+++ packages/v4/gatsby-config.js
@@ -35,7 +35,7 @@ module.exports = {
       options: {
         context: 'org',
         hiddenPages: ['withOuia'], // By title
-        showGdprBanner: false, // GDPR banner
+        showGdprBanner: true, // GDPR banner
         showFooter: true, // Footer
         sideNav: {
           core: [
@@ -126,7 +126,7 @@ module.exports = {
       resolve: 'gatsby-source-filesystem',
       options: {
         name: 'core', // This goes in URLs
-        path: require.resolve('@patternfly/patternfly/site/training.md')
+        path: path.join(__dirname, '/src/content/training/core-training.md')
       }
     },
     // React docs (MD + TSX)
@@ -190,7 +190,7 @@ module.exports = {
       resolve: 'gatsby-source-filesystem',
       options: {
         name: 'react', // This goes in URLs
-        path: require.resolve('@patternfly/react-docs/src/training.md')
+        path: path.join(__dirname, '/src/content/training/react-training.md')
       }
     },
     // Org docs