+++ packages/v4/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.32](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.31...patternfly-org-4@4.0.32) (2020-06-24)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.31](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.30...patternfly-org-4@4.0.31) (2020-06-22)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.30](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.29...patternfly-org-4@4.0.30) (2020-06-22)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.29](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.28...patternfly-org-4@4.0.29) (2020-06-22)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.28](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.27...patternfly-org-4@4.0.28) (2020-06-22)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.27](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.26...patternfly-org-4@4.0.27) (2020-06-11)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.26](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.25...patternfly-org-4@4.0.26) (2020-06-11)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.25](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.24...patternfly-org-4@4.0.25) (2020-06-11)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.24](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.23...patternfly-org-4@4.0.24) (2020-06-10)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.0.23](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.22...patternfly-org-4@4.0.23) (2020-06-09)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
 ## [4.0.22](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.0.21...patternfly-org-4@4.0.22) (2020-06-09)